Fixes following a Bazel upgrade on our end
We seem to be on a version of Bazel that now performs depth 1 checkouts
but that does not yet incorporate:

https://github.com/bazelbuild/bazel/commit/6496b1d1d25025f33406b1eaf9949cab126f08bd

that would presumably let us make our checkouts a bit less shallow, and
thus pin to arbitrary commits.  Instead, we must at the moment pin to
actual tags and branches on GitHub.  This diff makes the changes
necessary to pin to available versions instead of arbitrary commits;
tweaks a few things that broke during the necessary upgrades this
entailed; and where appropriate, follows the current Bazel docs in
replacing git repository fetches with HTTP archives.

The one thing I don’t like is specifying “master” for the repository
“org_pubref_rules_protobuf” — should we instead ask them to make a more
recent release?  Or fork the repository ourselves as a target?

Switch from “master” to permanent tag

Thanks to the upstream maintainer’s response to issue:
https://github.com/pubref/rules_protobuf/issues/199